Implement business rules for appointment module

+++ src/main/java/com/fabiana/mapa_viagem/service/AgendamentoService.java
@@ -2,12 +2,14 @@ package com.fabiana.mapa_viagem.service;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.fabiana.mapa_viagem.dto.AgendamentoDTO;
 import com.fabiana.mapa_viagem.exception.RecursoNaoEncontradoException;
+import com.fabiana.mapa_viagem.exception.RegraNegocioException;
 import com.fabiana.mapa_viagem.model.Acompanhante;
 import com.fabiana.mapa_viagem.model.Agendamento;
 import com.fabiana.mapa_viagem.model.Hospital;
@@ -38,6 +40,9 @@ public class AgendamentoService {
 	
 	@Autowired
 	private ViagemRepository viagemRepository;
+	
+	@Autowired
+	private ViagemService viagemService;
 	
 	
 	public List<AgendamentoDTO> findAll(){
@@ -71,22 +76,34 @@ public class AgendamentoService {
 		//Buscar Viagem
 		Viagem viagem =  viagemRepository.findById(agendamentoDto.getViagemId()).orElseThrow(() -> new RecursoNaoEncontradoException("Viagem não encontrada"));
 		
+		validarAgendamento(agendamentoDto, paciente, hospital, viagem);
+		
 		 Agendamento entity = fromDTO(agendamentoDto, paciente, acompanhante, hospital, viagem);
 		 entity = agendamentoRepository.save(entity);
 		 return new AgendamentoDTO(entity);
 	}
 	
 	 public void delete(Long id) {
-		 agendamentoRepository.findById(id).orElseThrow(() -> new RecursoNaoEncontradoException("Agendamento não encontrado"));
-		 agendamentoRepository.deleteById(id);
+	    Agendamento agendamento = agendamentoRepository.findById(id).orElseThrow(() -> new RecursoNaoEncontradoException("Agendamento não encontrado"));
+  	   //Verificando se a viagem foi iniciada  
+	    if (viagemService.viagemIniciada(agendamento.getViagem().getId())) 
+  	    	throw new RegraNegocioException("A viagem já foi iniciada, não é permitido deletar.");   
+	    agendamentoRepository.deleteById(id);	        
 	}
 	 
 	 @Transactional
 	 public AgendamentoDTO update(Long id, AgendamentoDTO dto) {
-
 		 Agendamento agendamento = agendamentoRepository.findById(id)
 	         .orElseThrow(() -> new RecursoNaoEncontradoException("Agendamento não encontrado"));
-		 
+
+		//Buscar Viagem
+		Viagem viagem =  viagemRepository.findById(dto.getViagemId()).orElseThrow(() -> new RecursoNaoEncontradoException("Viagem não encontrada"));
+			
+		//Verificar se a viagem já foi iniciada
+		if (viagemService.viagemIniciada(viagem.getId())) {
+		        throw new RegraNegocioException("A viagem já foi iniciada, não é permitido alterar.");
+		 }
+		
 		//Buscar Paciente
 		Paciente paciente =  pacienteRepository.findById(dto.getPacienteId()).orElseThrow(() -> new RecursoNaoEncontradoException("paciente não encontrado"));
 			
@@ -95,9 +112,9 @@ public class AgendamentoService {
 			
 		//Buscar Hospital
 		Hospital hospital =  hospitalRepository.findById(dto.getHospitalId()).orElseThrow(() -> new RecursoNaoEncontradoException("Hospital não encontrado"));
-
-		//Buscar Viagem
-		Viagem viagem =  viagemRepository.findById(dto.getViagemId()).orElseThrow(() -> new RecursoNaoEncontradoException("Viagem não encontrada"));
+		
+		//Validar regras de negocio
+		validarAgendamento(dto, paciente, hospital, viagem);
 		
 		 agendamento.setPaciente(paciente);
 		 agendamento.setAcompanhante(acompanhante);
@@ -114,6 +131,32 @@ public class AgendamentoService {
 			
 			
 	}
-	
+	 
+	//validar regras antes de inserir ou alterar 
+	public void validarAgendamento (AgendamentoDTO objDto, Paciente paciente, Hospital hospital, Viagem viagem) {
+		
+		if (!objDto.getDataAtendimento().equals(viagem.getDataViagem())) {
+			 throw new RegraNegocioException("Data do atendimento diferente da data de viagem.");
+		}
+		
+		if (!hospital.getCidade().trim().equalsIgnoreCase(viagem.getCidadeDestino().trim())) {
+		  throw new RegraNegocioException("A Cidade destino é diferente da cidade onde está localizado o hospital.");
+		}
+		
+		if (objDto.getHorarioAtendimento().isBefore(viagem.getHoraPrevista())) {
+		 throw new RegraNegocioException("Horario de antedimento anterior a hora prevista da viagem.");	
+		}
+					   
+	    // Verifica duplicidade de agendamento
+        Optional<Agendamento> agendamentoExistente = agendamentoRepository.findByPacienteIdAndViagemId(paciente.getId(), viagem.getId());
+
+        if (agendamentoExistente.isPresent() &&
+        	    (objDto.getId() == null || !agendamentoExistente.get().getId().equals(objDto.getId()))) {
+        	    throw new RegraNegocioException("Paciente já possui agendamento nesta viagem.");
+        }
+	}
+	 
+
+
 
 }
